Add files via upload
This repository contains a jar file that demonstrates two things: 
i. A static method of a class can be called "on" an instance of that class, using the syntax normally used for calling instance methods.
ii. A static method is "statically resolved" at compile-time. This is as opposed to a non-static method, which is "dynamically resolved" at runtime, which is the heart of polymorphism in Java and similar languages!